+++ internal/pkg/instrumentation/bpf/net/http/server/bpf_bpfel_arm64.go
@@ -28,6 +28,9 @@ type bpfUprobeDataT struct {
 		StatusCode uint64
 		Method     [8]int8
 		Path       [128]int8
+		RemoteAddr [256]int8
+		Host       [256]int8
+		Proto      [8]int8
 	}
 	RespPtr uint64
 }